Changes to vim config, primarily removing Vundle for plugin management
A lot of cosmetic changes were made to vimrc, but the most major change
is adding Rake tasks to manage plugins for vim. Rather than embed them
in the vimrc, plugins are now listed in vim/plugins, which is read by
Rake to determine what to install or remove. I liked how Vundle
organized it though, so I've tried to keep a similar layout between
generic Github repos, and repos under the Vim-Scripts account.